Update example and add one, multiple bug fixes

+++ include/Intersectables/TriangleMeshObject.h
@@ -34,7 +34,7 @@ class TriangleMeshObject : public IIntersectable
      * @param indices   Vector of indices for the vertices. Every 3 indices define one triangle.
      * @param material  The objects material.
      */
-    TriangleMeshObject(std::vector<Vertex> vertices, std::vector<std::uint64_t> indices, Material material);
+    TriangleMeshObject(std::vector<Vertex> vertices, std::vector<std::uint32_t> indices, Material material);
 
     /**
      * Destructor, cleans up this object on deletion.
@@ -94,16 +94,6 @@ class TriangleMeshObject : public IIntersectable
     bool operator!=(const IIntersectable& object) const override;
 
   private:
-    // friend class Triangle;
-    //
-    // std::vector<Vertex>   vertices;
-    // std::vector<uint64_t> indices;
-    // Material              material;
-    //
-    // std::vector<std::unique_ptr<IIntersectable>> triangles;
-    // DBVHv2                                       structure;
-
-    // TODO
     class TrianglMeshImpl;
 
     std::unique_ptr<TrianglMeshImpl> impl_;
